+++ src/options/pages/blueprints/useInstallableViewItems.tsx
@@ -19,8 +19,6 @@ import {
   Installable,
   InstallableStatus,
   InstallableViewItem,
-  SharingSource,
-  SharingType,
 } from "@/options/pages/blueprints/blueprintsTypes";
 import React, { useCallback, useMemo } from "react";
 import { useSelector } from "react-redux";
@@ -31,12 +29,11 @@ import {
   getDescription,
   getInstalledVersionNumber,
   getLabel,
-  getOrganization,
   getPackageId,
+  getSharingType,
   getUpdatedAt,
+  isDeployment,
   isExtension,
-  isPersonal,
-  isPublic,
   updateAvailable,
 } from "@/options/pages/blueprints/utils/installableUtils";
 import {
@@ -81,65 +78,9 @@ function useInstallableViewItems(installables: Installable[]): {
     [installedExtensionIds, installedRecipeIds]
   );
 
-  const isDeployment = useCallback(
-    (installable: Installable) => {
-      if (isExtension(installable)) {
-        return Boolean(installable._deployment);
-      }
-
-      for (const installedExtension of installedExtensions) {
-        if (
-          installedExtension._recipe?.id === getPackageId(installable) &&
-          installedExtension._deployment
-        ) {
-          return true;
-        }
-      }
-
-      return false;
-    },
-    [installedExtensions]
-  );
-
-  const getSharingType = useCallback(
-    (installable: Installable, scope: string): SharingSource => {
-      let sharingType: SharingType = null;
-      const organization = getOrganization(installable, organizations);
-
-      if (isPersonal(installable, scope)) {
-        sharingType = "Personal";
-      } else if (isDeployment(installable)) {
-        sharingType = "Deployment";
-      } else if (organization) {
-        sharingType = "Team";
-      } else if (isPublic(installable)) {
-        sharingType = "Public";
-      }
-
-      let label: string;
-      if (
-        sharingType === "Team" ||
-        // There's a corner case for team deployments of public market bricks. The organization will come through as
-        // nullish here.
-        (sharingType === "Deployment" && organization?.name)
-      ) {
-        label = organization.name;
-      } else {
-        label = sharingType;
-      }
-
-      return {
-        type: sharingType,
-        label,
-        organization,
-      };
-    },
-    [isDeployment, organizations]
-  );
-
   const getStatus = useCallback(
     (installable: Installable): InstallableStatus => {
-      if (isDeployment(installable)) {
+      if (isDeployment(installable, installedExtensions)) {
         if (isExtension(installable)) {
           return isDeploymentActive(installable) ? "Active" : "Paused";
         }
@@ -156,7 +97,7 @@ function useInstallableViewItems(installables: Installable[]): {
 
       return isActive(installable) ? "Active" : "Inactive";
     },
-    [installedExtensions, isActive, isDeployment]
+    [installedExtensions, isActive]
   );
 
   const installableIcon = useCallback(
@@ -184,7 +125,12 @@ function useInstallableViewItems(installables: Installable[]): {
         description: getDescription(installable),
         sharing: {
           packageId: getPackageId(installable),
-          source: getSharingType(installable, scope),
+          source: getSharingType(
+            installable,
+            organizations,
+            scope,
+            installedExtensions
+          ),
         },
         updatedAt: getUpdatedAt(installable),
         status: getStatus(installable),
